Moved is_intersect_square comment_line's logic
