+++ src/canvas.rs
@@ -472,76 +472,6 @@ impl Canvas {
         }
     }
 
-    /// Orientation of a, b, c.
-    /// - `0`  - collinear (lie on the same line)
-    /// - `1`  - clockwise rotation
-    /// - `-1` - counterclockwise rotation
-    fn orientation(&self, a: Pos2, b: Pos2, c: Pos2) -> i32 {
-        let value = (b.y - a.y) * (c.x - a.x) - (b.x - a.x) * (c.y - b.y);
-        value.signum() as i32
-    }
-
-    /// Check if b is on (a; c) segment
-    fn on_segment(&self, a: Pos2, b: Pos2, c: Pos2) -> bool {
-        b.x >= a.x.min(c.x) && b.x <= a.x.max(c.x) && b.y >= a.y.min(c.y) && b.y <= a.y.max(c.y)
-    }
-
-    /// Check if segments (a; b) and (c; d) are intersecting
-    fn segments_intersect(&self, a: Pos2, b: Pos2, c: Pos2, d: Pos2) -> bool {
-        let o1 = self.orientation(a, b, c);
-        let o2 = self.orientation(a, b, d);
-        let o3 = self.orientation(c, d, a);
-        let o4 = self.orientation(c, d, b);
-
-        if o1 != o2 && o3 != o4 {
-            return true;
-        }
-
-        (o1 == 0 && self.on_segment(a, c, b))
-            || (o2 == 0 && self.on_segment(a, d, b))
-            || (o3 == 0 && self.on_segment(c, a, d))
-            || (o4 == 0 && self.on_segment(c, b, d))
-    }
-
-    /// Check if `comment_line` is intersect line
-    fn is_intersect(&self, comment_line: &CommentLine, line: [Pos2; 2]) -> bool {
-        let c = line[0];
-        let d = line[1];
-
-        for pair in comment_line.points.windows(2) {
-            let a = pair[0];
-            let b = pair[1];
-
-            if self.segments_intersect(a, b, c, d) {
-                return true;
-            }
-        }
-
-        false
-    }
-
-    /// Check if square is intersect `comment_line`
-    pub fn is_intersect_square(&self, comment_line: &CommentLine, square: Rect) -> bool {
-        let square_edges = [
-            [square.left_top(), square.right_top()],
-            [square.right_top(), square.right_bottom()],
-            [square.right_bottom(), square.left_bottom()],
-            [square.left_bottom(), square.left_top()],
-        ];
-
-        for point in &comment_line.points {
-            if square.contains(*point)
-                || square_edges
-                    .iter()
-                    .any(|edge| self.is_intersect(comment_line, *edge))
-            {
-                return true;
-            }
-        }
-
-        false
-    }
-
     pub fn handle_comment_erase(&mut self, comment_lines: &mut CommentsGroup) {
         if self.response().hover_pos().is_none() {
             return;
@@ -566,7 +496,7 @@ impl Canvas {
             // Find comment_line intersected by interact_square
             let mut selected_line_id = None;
             for (id, line) in comment_lines.iter() {
-                if self.is_intersect_square(line, interact_square) {
+                if line.is_intersect_square(interact_square) {
                     selected_line_id = Some(id);
                     break;
                 }
